### 2022-11-27 #4976 @editorialbot
[PRE REVIEW]: mlpack 4: a fast, header-only C++ machine learning library
**Submitting author:** @rcurtin (<a href="http://orcid.org/0000-0002-9903-8214">Ryan Curtin</a>)
**Repository:** https://github.com/mlpack/mlpack
**Branch with paper.md** (empty if default branch): mlpack4_joss
**Version:** v4.0.0
