### 2022-12-01 #4976 @conradsnicta
editorialbot wrote:
> Failed to discover a `Statement of need` section in paper

Strange. There was always a "Statement of Need" section on the first page of the paper; see line 24 within the generated pdf.
