### 2022-12-01 #4976 @rcurtin
Yeah, it picked up a different paper (the mlpack 3 paper) in the repository, which I'd forgotten to remove.  I pushed a commit to the branch and now it renders the right one, where there is a Statement of Need section.
