### 2022-12-01 c6f95be
Only add Python test if BUILD_TESTS is ON.
