### 2022-12-01 ce655c1
Fix TMATRIX handling for Python by using doTranspose member.
