### 2022-12-01 1e445ca
Add transpose matrix support for Julia bindings.
