+++ src/mlpack/bindings/julia/print_input_processing_impl.hpp
@@ -99,7 +99,12 @@ void PrintInputProcessing(
   else
   {
     matTypeModifier = "Mat";
-    extra = ", points_are_rows";
+
+    // Determine whether we need to transpose the matrix.
+    std::string transStr =
+        (d.noTranspose ? std::string("true") : std::string("false"));
+
+    extra = ", points_are_rows, " + transStr;
   }
 
   // Now print the SetParam call.
