+++ src/mlpack/bindings/julia/mlpack/params.jl.in
@@ -47,12 +47,14 @@ function convert_to_1d(in::Array{T, 2})::Array{T, 1} where T
 end
 
 # Utility function to convert to and return a matrix.
-function to_matrix(input, T::Type)
-  if isa(input, Array{T, 1})
+function to_matrix(input, T::Type, transpose::Bool)
+  result = if isa(input, Array{T, 1})
     convert_to_2d(input)
   else
     convert(Array{T, 2}, input)
   end
+
+  return transpose ? convert(Array{T, 2}, result') : result
 end
 
 # Utility function to convert to and return a vector.
@@ -104,9 +106,10 @@ function SetParamMat(params::Ptr{Nothing},
                      paramName::String,
                      paramValue,
                      pointsAsRows::Bool,
+                     transpose::Bool,
                      juliaOwnedMemory::Set{Ptr{Nothing}})
   push!(juliaOwnedMemory, convert(Ptr{Nothing}, Base.pointer(paramValue)))
-  paramMat = to_matrix(paramValue, Float64)
+  paramMat = to_matrix(paramValue, Float64, transpose)
   ccall((:SetParamMat, library), Nothing, (Ptr{Nothing}, Cstring, Ptr{Float64},
       Csize_t, Csize_t, Bool), params, paramName, Base.pointer(paramMat),
       size(paramMat, 1), size(paramMat, 2), pointsAsRows)
@@ -116,9 +119,10 @@ function SetParamUMat(params::Ptr{Nothing},
                       paramName::String,
                       paramValue,
                       pointsAsRows::Bool,
+                      transpose::Bool,
                       juliaOwnedMemory::Set{Ptr{Nothing}})
   push!(juliaOwnedMemory, convert(Ptr{Nothing}, Base.pointer(paramValue)))
-  paramMat = to_matrix(paramValue, Int)
+  paramMat = to_matrix(paramValue, Int, transpose)
 
   # Sanity check.
   if minimum(paramMat) <= 0
