### 2022-12-01 8b5db73
Handle transposed matrices for Go.
