### 2022-12-01 ec704c1
Fix transpose matrix handling for R bindings.
