### 2022-12-02 5758547
Update missed examples to the mlpack 4 API.
