### 2022-12-02 e028738
Split image into separate file.
