+++ doc/tutorials/ann.md
@@ -104,70 +104,7 @@ the [matrices guide](../user/matrices.md) for more information.
 The code below builds a simple feed-forward network with the default options,
 then queries for the assignments for every point in the `queries` matrix.
 
-TODO: move to separate rendered image
-\dot
-digraph G {
-  fontname = "Hilda 10"
-  rankdir=LR
-  splines=line
-  nodesep=.08;
-  ranksep=1;
-  edge [color=black, arrowsize=.5];
-  node [fixedsize=true,label="",style=filled,color=none,fillcolor=gray,shape=circle]
-
-  subgraph cluster_0 {
-    color=none;
-    node [style=filled, color=white, penwidth=15,fillcolor=black shape=circle];
-    l10  l11  l12  l13  l14  l15  ;
-    label = Input;
-  }
-
-  subgraph cluster_1 {
-    color=none;
-    node [style=filled, color=white, penwidth=15,fillcolor=gray shape=circle];
-    l20  l21  l22  l23  l24  l25  l26  l27  ;
-    label = Linear;
-  }
-
-  subgraph cluster_2 {
-    color=none;
-    node [style=filled, color=white, penwidth=15,fillcolor=gray shape=circle];
-    l30  l31  l32  l33  l34  l35  l36  l37  ;
-    label = Linear;
-  }
-
-  subgraph cluster_3 {
-    color=none;
-    node [style=filled, color=white, penwidth=15,fillcolor=black shape=circle];
-    l40  l41  l42  ;
-    label = LogSoftMax;
-  }
-
-  l10 -> l20   l10 -> l21   l10 -> l22   l10 -> l23   l10 -> l24   l10 -> l25
-  l10 -> l26   l10 -> l27   l11 -> l20   l11 -> l21   l11 -> l22   l11 -> l23
-  l11 -> l24   l11 -> l25   l11 -> l26   l11 -> l27   l12 -> l20   l12 -> l21
-  l12 -> l22   l12 -> l23   l12 -> l24   l12 -> l25   l12 -> l26   l12 -> l27
-  l13 -> l20   l13 -> l21   l13 -> l22   l13 -> l23   l13 -> l24   l13 -> l25
-  l13 -> l26   l13 -> l27   l14 -> l20   l14 -> l21   l14 -> l22   l14 -> l23
-  l14 -> l24   l14 -> l25   l14 -> l26   l14 -> l27   l15 -> l20   l15 -> l21
-  l15 -> l22   l15 -> l23   l15 -> l24   l15 -> l25   l15 -> l26   l15 -> l27
-  l20 -> l30   l20 -> l31   l20 -> l32   l20 -> l33   l20 -> l34   l20 -> l35
-  l20 -> l36   l20 -> l37   l21 -> l30   l21 -> l31   l21 -> l32   l21 -> l33
-  l21 -> l34   l21 -> l35   l21 -> l36   l21 -> l37   l22 -> l30   l22 -> l31
-  l22 -> l32   l22 -> l33   l22 -> l34   l22 -> l35   l22 -> l36   l22 -> l37
-  l23 -> l30   l23 -> l31   l23 -> l32   l23 -> l33   l23 -> l34   l23 -> l35
-  l23 -> l36   l23 -> l37   l24 -> l30   l24 -> l31   l24 -> l32   l24 -> l33
-  l24 -> l34   l24 -> l35   l24 -> l36   l24 -> l37   l25 -> l30   l25 -> l31
-  l25 -> l32   l25 -> l33   l25 -> l34   l25 -> l35   l25 -> l36   l25 -> l37
-  l26 -> l30   l26 -> l31   l26 -> l32   l26 -> l33   l26 -> l34   l26 -> l35
[…]
