### 2022-12-02 1b4b77b
Move files that aren't docs into their own folder.
