### 2022-12-02 4b83860
Remove last references to doxygen in the documentation.
