+++ .github/workflows/update-catch.yaml
@@ -11,7 +11,7 @@ on:
     contents: read
 jobs:
   updateCatch:
-    permissions
+    permissions:
       contents: write # for peter-evans/create-pull-request to create branch
       pull-requests: write # for peter-evans/create-pull-request to create a PR
     if: ${{ false }}
