+++ .github/workflows/update-catch.yaml
@@ -7,8 +7,8 @@ on:
   workflow_dispatch:
   schedule:
     - cron:  '0 10 1/16 * *'
-  permissions:
-    contents: read
+permissions:
+  contents: read
 
 jobs:
   updateCatch:
