### 2022-12-06 67493c2
remove second "Quick Links" to reduce confusion
remove second "Quick Links" to reduce confusion with first "Quick Links" on line 40
