### 2022-12-06 a64043f
more fixes due to markdown limitations
