### 2022-12-08 a080b40
input to input_ in python quickstart
