### 2022-12-15 c96b062
add missing angle brackets in cpp simple quickstart example
