### 2022-12-20 #4976 @osorensen
:wave: @sandeep-ps, @Radonirinaunimi, @max-little, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
