### 2022-12-21 78d10cc
implementing the concept of border pointsand fixing error
