+++ src/mlpack/methods/dbscan/dbscan_impl.hpp
@@ -102,23 +102,61 @@ size_t DBSCAN<RangeSearchType, PointSelectionPolicy>::Cluster(
   // Initialize the UnionFind object.
   UnionFind uf(data.n_cols);
   rangeSearch.Train(data);
-
-  if (batchMode)
-    BatchCluster(data, uf);
+  std::vector<std::vector<size_t>> neighbors;
+  if (batchMode){
+    BatchCluster(data, uf, neighbors);
+    }
   else
     PointwiseCluster(data, uf);
 
-  // Now set assignments.
+  // Now set assignments of core points.
   assignments.set_size(data.n_cols);
   for (size_t i = 0; i < data.n_cols; ++i)
     assignments[i] = uf.Find(i);
 
-  // Get a count of all clusters.
+  // Get a count of all clusters of core points.
   const size_t numClusters = arma::max(assignments) + 1;
   arma::Col<size_t> counts(numClusters, arma::fill::zeros);
   for (size_t i = 0; i < assignments.n_elem; ++i)
     counts[assignments[i]]++;
-
+    
+  /**
+  * Performs addition of border points to clusters
+  * Also updating the list of counts and assign
+  */
+  if (batchMode){
+    for(size_t i=0; i < data.n_cols;++i){
+      if(counts[i]<minPoints && assignments[i]==i){
+        for(size_t j=0;j<neighbors[i].size();j++){
+          if (counts[assignments[neighbors[i][j]]]>=minPoints){
+	    uf.Union(i, neighbors[i][j]);
+	    assignments[i]=assignments[neighbors[i][j]];
+	    counts[assignments[neighbors[i][j]]]++;
+	    counts[i]--;
+	    break;
+	  }
+        }
+      }
+    }
+  }
+  else{
+    for(size_t i=0; i < data.n_cols;++i){
+      if(counts[i]<minPoints && assignments[i]==i){
+        std::vector<std::vector<double>> distances;
+        rangeSearch.Search(data.col(i), Range(0.0, epsilon),
+          neighbors, distances);
+        for(size_t j=0;j<neighbors[0].size();j++){
+	  if (counts[assignments[neighbors[0][j]]]>=minPoints){
+	    uf.Union(i, neighbors[0][j]);
+	    assignments[i]=assignments[neighbors[0][j]];
+	    counts[assignments[neighbors[0][j]]]++;
+	    counts[i]--;
+	    break;
+	  }
+	}
+      }
+    }
+  }
   // Now assign clusters to new indices.
   size_t currentCluster = 0;
   arma::Col<size_t> newAssignments(numClusters);
@@ -153,7 +191,9 @@ void DBSCAN<RangeSearchType, PointSelectionPolicy>::PointwiseCluster(
 {
   std::vector<std::vector<size_t>> neighbors;
   std::vector<std::vector<double>> distances;
-
+  // marking core points in order to cluster them
+  arma::Row<size_t> core(data.n_cols);
+  core.fill(0);
   for (size_t i = 0; i < data.n_cols; ++i)
   {
     if (i % 10000 == 0 && i > 0)
@@ -162,10 +202,12 @@ void DBSCAN<RangeSearchType, PointSelectionPolicy>::PointwiseCluster(
     // Do the range search for only this point.
     rangeSearch.Search(data.col(i), Range(0.0, epsilon), neighbors,
         distances);
-
-    // Union to all neighbors.
+    if(neighbors[0].size()>=minPoints) core[i]=1;
+    else continue;
+    // Union to all core points with their neighbor core points
     for (size_t j = 0; j < neighbors[0].size(); ++j)
-      uf.Union(i, neighbors[0][j]);
+      if(core[neighbors[0][j]])
+        uf.Union(i, neighbors[0][j]);
[…]
