+++ src/mlpack/methods/dbscan/dbscan.hpp
@@ -150,7 +150,8 @@ class DBSCAN
    * @param uf UnionFind structure that will be modified.
    */
   template<typename MatType>
-  void BatchCluster(const MatType& data, UnionFind& uf);
+  void BatchCluster(const MatType& data, UnionFind& uf,
+  std::vector<std::vector<size_t>>& neighbors);
 };
 
 } // namespace mlpack
