+++ CMake/src/mlpack/methods/CMakeFiles/mlpack_range_search.dir/link.txt
@@ -1,0 +1,1 @@
+/usr/bin/c++   -Wall -Wextra -ftemplate-depth=1000 -Wno-unused-function -O3 -fopenmp   CMakeFiles/mlpack_range_search.dir/range_search/range_search_main.cpp.o  -o ../../../bin/mlpack_range_search  -lpthread /usr/lib/libarmadillo.so /usr/lib/gcc/x86_64-linux-gnu/9/libgomp.so /usr/lib/x86_64-linux-gnu/libpthread.so 
