+++ CMake/CMakeCache.txt
@@ -1,727 +1,0 @@
-# This is the CMakeCache file.
-# For build in directory: /home/adarsh/gsoc/mlpack/CMake
-# It was generated by CMake: /usr/bin/cmake
-# You can edit this file to change values found and used by cmake.
-# If you do not want to change any of the values, simply exit the editor.
-# If you do want to change a value, simply edit, save, and exit the editor.
-# The syntax for the file is as follows:
-# KEY:TYPE=VALUE
-# KEY is the name of a variable in the cache.
-# TYPE is a hint to GUIs for the type of VALUE, DO NOT EDIT TYPE!.
-# VALUE is the current value for the KEY.
-
-########################
-# EXTERNAL cache entries
-########################
-
-//Path to a file.
-ARMADILLO_INCLUDE_DIR:PATH=/home/ryan/armadillo/include
-
-//No help, variable specified on the command line.
-ARMADILLO_LIBRARY:UNINITIALIZED=/home/ryan/armadillo/lib/libarmadillo.so
-
-//Compile with extra Armadillo debugging symbols.
-ARMA_EXTRA_DEBUG:BOOL=OFF
-
-//Build command-line executables.
-BUILD_CLI_EXECUTABLES:BOOL=ON
-
-//Build Go bindings.
-BUILD_GO_BINDINGS:BOOL=OFF
-
-//Build Go shared library.
-BUILD_GO_SHLIB:BOOL=OFF
-
-//Build Julia bindings.
-BUILD_JULIA_BINDINGS:BOOL=OFF
-
-//Build Markdown bindings for website documentation.
-BUILD_MARKDOWN_BINDINGS:BOOL=OFF
-
-//Build Python bindings.
-BUILD_PYTHON_BINDINGS:BOOL=OFF
-
-//Build R bindings.
-BUILD_R_BINDINGS:BOOL=OFF
-
-//Compile shared objects for tests and bindings (if OFF, static
-// libraries and binaries are compiled).
-BUILD_SHARED_LIBS:BOOL=ON
-
-//Build the testing tree.
-BUILD_TESTING:BOOL=ON
-
-//Build tests.
-BUILD_TESTS:BOOL=ON
-
-//Path to a program.
-BZRCOMMAND:FILEPATH=BZRCOMMAND-NOTFOUND
-
-//Path to a file.
-CEREAL_INCLUDE_DIR:PATH=/usr/local/include
-
-//Path to a program.
-CMAKE_ADDR2LINE:FILEPATH=/usr/bin/addr2line
-
-//Path to a program.
-CMAKE_AR:FILEPATH=/usr/bin/ar
-
-//Choose the type of build, options are: None Debug Release RelWithDebInfo
-// MinSizeRel ...
-CMAKE_BUILD_TYPE:STRING=
-
-//Enable/Disable color output during build.
-CMAKE_COLOR_MAKEFILE:BOOL=ON
-
-//CXX compiler
-CMAKE_CXX_COMPILER:FILEPATH=/usr/bin/c++
-
-//A wrapper around 'ar' adding the appropriate '--plugin' option
-// for the GCC compiler
-CMAKE_CXX_COMPILER_AR:FILEPATH=/usr/bin/gcc-ar-9
-
-//A wrapper around 'ranlib' adding the appropriate '--plugin' option
-// for the GCC compiler
-CMAKE_CXX_COMPILER_RANLIB:FILEPATH=/usr/bin/gcc-ranlib-9
-
[…]
