+++ CMake/src/mlpack/methods/CMakeFiles/mlpack_mean_shift.dir/link.txt
@@ -1,1 +1,0 @@
-/usr/bin/c++   -Wall -Wextra -ftemplate-depth=1000 -Wno-unused-function -O3 -fopenmp   CMakeFiles/mlpack_mean_shift.dir/mean_shift/mean_shift_main.cpp.o  -o ../../../bin/mlpack_mean_shift  -lpthread /usr/lib/libarmadillo.so /usr/lib/gcc/x86_64-linux-gnu/9/libgomp.so /usr/lib/x86_64-linux-gnu/libpthread.so 
