### 2022-12-22 #4976 @zhangjy-ge
> 👋 @zhangjy-ge, @AP6YC, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi, yes I would be happy to review paper.
