### 2022-12-23 bc4ae68
comment out add_definitions(-DARMA_NO_DEBUG)
Co-authored-by: Ryan Curtin <ryan@ratml.org>
