+++ CMakeLists.txt
@@ -246,7 +246,6 @@ if (DEBUG)
     endif()
   endif()
 else()
-  # add_definitions(-DARMA_NO_DEBUG)
   add_definitions(-DNDEBUG)
   if (NOT MSVC)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3")
