### 2022-12-23 e66b93c
Add new block for next release to HISTORY.md.
