+++ HISTORY.md
@@ -1,3 +1,6 @@
+### mlpack ?.?.?
+###### ????-??-??
+
 ### mlpack 4.0.1
 ###### 2022-12-23
   * Fix mapping of categorical data for Julia bindings (#3305).
