+++ CMakeLists.txt
@@ -339,7 +339,7 @@ if (NOT DOWNLOAD_DEPENDENCIES)
 else()
   find_package(Ensmallen "${ENSMALLEN_VERSION}")
   if (NOT ENSMALLEN_FOUND)
-    get_deps(http://www.ensmallen.org/files/ensmallen-latest.tar.gz ensmallen ensmallen-latest.tar.gz)
+    get_deps(http://www.ensmallen.org/files/ensmallen-2.19.0.tar.gz ensmallen ensmallen-latest.tar.gz)
     set(ENSMALLEN_INCLUDE_DIR ${GENERIC_INCLUDE_DIR})
     find_package(Ensmallen REQUIRED)
   endif()
