### 2022-12-23 1b5d1c1
Add new block for next release to HISTORY.md.
