+++ HISTORY.md
@@ -1,5 +1,5 @@
-### mlpack ?.?.?
-###### ????-??-??
+### mlpack 4.0.1
+###### 2022-12-23
   * Fix mapping of categorical data for Julia bindings (#3305).
 
   * Bugfix: catch all exceptions when running bindings from Julia, instead of
