### 2022-12-27 #5026 @zhangjy-ge
I have finished my initial review. I think overall, mlpack 4 is a well written package that has indicated great potential in the machine learning field, considering its good documentation, user friendliness, and the algorithms it implemented, etc. Thus, I would recommend it for publication after minor revision:
1. even though I understand that in mlpack's website, there are some comparisons between mlpack and other similar packages, it is not mentioned in the manuscript itself. I would suggest the authors to add such descriptions, based on the requirements of "State of the field" and "performance" in the checklist.
