### 2022-12-28 954b7a6
Added Size checks for Decision Tree, GMM, NCA, Random Forest
