+++ src/mlpack/methods/random_forest/random_forest_impl.hpp
@@ -22,16 +22,14 @@ template<
     typename DimensionSelectionType,
     template<typename> class NumericSplitType,
     template<typename> class CategoricalSplitType,
-    bool UseBootstrap,
-    typename MatType
+    bool UseBootstrap
 >
 RandomForest<
     FitnessFunction,
     DimensionSelectionType,
     NumericSplitType,
     CategoricalSplitType,
-    UseBootstrap,
-    MatType
+    UseBootstrap
 >::RandomForest() :
     avgGain(0.0)
 {
@@ -43,16 +41,15 @@ template<
     typename DimensionSelectionType,
     template<typename> class NumericSplitType,
     template<typename> class CategoricalSplitType,
-    bool UseBootstrap,
-    typename MatType
+    bool UseBootstrap
 >
+template<typename MatType>
 RandomForest<
     FitnessFunction,
     DimensionSelectionType,
     NumericSplitType,
     CategoricalSplitType,
-    UseBootstrap,
-    MatType
+    UseBootstrap
 >::RandomForest(const MatType& dataset,
                 const arma::Row<size_t>& labels,
                 const size_t numClasses,
@@ -61,9 +58,8 @@ RandomForest<
                 const double minimumGainSplit,
                 const size_t maximumDepth,
                 DimensionSelectionType dimensionSelector) :
-    avgGain(0.0),
-    dataset(dataset)
-{ 
+    avgGain(0.0)
+{
   // Pass off work to the Train() method.
   data::DatasetInfo info; // Ignored.
   arma::rowvec weights; // Fake weights, not used.
@@ -77,16 +73,15 @@ template<
     typename DimensionSelectionType,
     template<typename> class NumericSplitType,
     template<typename> class CategoricalSplitType,
-    bool UseBootstrap,
-    typename MatType
+    bool UseBootstrap
 >
+template<typename MatType>
 RandomForest<
     FitnessFunction,
     DimensionSelectionType,
     NumericSplitType,
     CategoricalSplitType,
-    UseBootstrap,
-    MatType
+    UseBootstrap
 >::RandomForest(const MatType& dataset,
                 const data::DatasetInfo& datasetInfo,
                 const arma::Row<size_t>& labels,
@@ -96,8 +91,7 @@ RandomForest<
                 const double minimumGainSplit,
                 const size_t maximumDepth,
                 DimensionSelectionType dimensionSelector):
-                    avgGain(0.0),
-                    dataset(dataset)
+                    avgGain(0.0)
 {
   // Pass off work to the Train() method.
   arma::rowvec weights; // Fake weights, not used.
@@ -111,16 +105,15 @@ template<
     typename DimensionSelectionType,
     template<typename> class NumericSplitType,
     template<typename> class CategoricalSplitType,
-    bool UseBootstrap,
-    typename MatType
+    bool UseBootstrap
 >
+template<typename MatType>
 RandomForest<
     FitnessFunction,
     DimensionSelectionType,
     NumericSplitType,
     CategoricalSplitType,
-    UseBootstrap,
-    MatType
+    UseBootstrap
 >::RandomForest(const MatType& dataset,
                 const arma::Row<size_t>& labels,
                 const size_t numClasses,
@@ -130,8 +123,7 @@ RandomForest<
                 const double minimumGainSplit,
                 const size_t maximumDepth,
                 DimensionSelectionType dimensionSelector) :
-    avgGain(0.0),
-    dataset(dataset)
+    avgGain(0.0)
 {
   // Pass off work to the Train() method.
   data::DatasetInfo info; // Ignored by Train().
@@ -145,16 +137,15 @@ template<
[…]
