### 2022-12-29 f626041
Install libcereal-dev (and thus rapidjson-dev) for the R CMD check call.
