### 2022-12-31 #5026 @osorensen
> Hi, @osorensen,
>
> Two authors listed in the paper (#5 and #9) are affiliated with the same institution I'm affiliated with. We are from different units and haven't collaborated in the past, and we do not know each other. I request that this conflict be waived. Please let me know if there is a different process for this. Thanks.

Thanks for letting me know @sandeep-ps. I think this should be waived, but I'll ping the editor-in-chief @openjournals/dsais-eics to hear what they think.
