### 2022-12-31 #5026 @gkthiruvathukal
@sandeep-ps Thanks for letting us know. I'm a bit reticent about waiving the COI because there are implications of reviewing work by others at your institution if you are both still employed there. I'd prefer that we seek a different reviewer and hope we can assign you to another submission where we need reviewers. Ok? Thank you for understanding.

@osorensen Please assign to a different reviewer and keep @sandeep-ps in mind for other assignments!
