### 2023-01-05 c8d7a95
Fix unused UnionFind point.
