### 2023-01-05 8e1f30b
Fix definition and usage of minPoints.
