### 2023-01-05 9a7fb84
Attempt to install cereal on OS X runner.
