### 2023-01-05 496c6ac
Add very simple attempt at Markdown documentation for decision trees.
