+++ doc/user/methods/decision_tree.md
@@ -1,0 +1,135 @@
+## `DecisionTree`
+
+The `DecisionTree` class implements a decision tree classifier that supports
+numerical and categorical features, by default using Gini gain to choose which
+feature to split on.  The class offers several template parameters and several
+constructor parameters that can be used to control the behavior of the tree.
+
+### Constructors
+
+ * `DecisionTree(`_`numClasses`_`)` (initialize without training)
+
+ * `DecisionTree(`_`data`_`, `_`labels`_`, `_`numClasses`_`)` (train on
+   numerical-only data)
+ * `DecisionTree(`_`data`_`, `_`labels`_`, `_`numClasses`_`, `_`minimumLeafSize`_`, `_`minimumGainSplit`_`, `_`maximumDepth`_`)` (train on numerical-only data with hyperparameters)
+
+<!-- TODO: weighted numerical-only constructors -->
+
+ * `DecisionTree(`_`data`_`, `_`datasetInfo`_`, `_`labels`_`, `_`numClasses`_`)` (train on mixed categorical data)
+ * `DecisionTree(`_`data`_`, `_`datasetInfo`_`, `_`labels`_`, `_`numClasses`_`, `_`minimumLeafSize`_`, `_`minimumGainSplit`_`, `_`maximumDepth`_`)` (train on mixed categorical data with hyperparameters)
+
+ * `DecisionTree(`_`data`_`, `_`datasetInfo`_`, `_`labels`_`, `_`numClasses`_`,    `_`weights`_`)` (train on weighted mixed categorical data)
+ * `DecisionTree(`_`data`_`, `_`datasetInfo`_`, `_`labels`_`, `_`numClasses`_`, `_`weights`_`, `_`minimumLeafSize`_`, `_`minimumGainSplit`_`, `_`maximumDepth`_`)` (train on weighted mixed categorical data with hyperparameters)
+
+| **type** | **name** | **description** | **default** |
+|----------|----------|-----------------|-------------|
+| [`arma::mat`](../matrices.md) | **`data`** | [Column-major](../matrices.md) training matrix. | _(N/A)_ |
+<!-- TODO: the linked tutorial is inappropriate, and we need a simpler
+DatasetInfo tutorial that discusses how to load and work with categorical data
+in very straightforward terms -->
+| [`data::DatasetInfo`](../../tutorials/datasetmapper.md) | **`datasetInfo`** |
+Dataset information, specifying type information for each dimension. | _(N/A)_ |
+<!-- TODO: update matrices.md to include a section on labels, and how to use
+`NormalizeLabels()` to map to [0, numClasses - 1]. -->
+| [`arma::Row<size_t>`]('../matrices.md') | **`labels`** | Training labels, between `0` and `numClasses - 1` (inclusive). | _(N/A)_ |
+<!-- TODO: include a bit about instance weights in matrices.md -->
+| [`arma::rowvec`]('../matrices.md') | **`weights`** | Weights for each training point. | _(N/A)_ |
+| `size_t` | **`numClasses`** | Number of classes in the dataset. | _(N/A)_ |
+| `size_t` | **`minimumLeafSize`** | Minimum number of points in each leaf node. | `10` |
+| `double` | **`minimumGainSplit`** | Minimum gain for a node to split. | `1e-7` |
+| `size_t` | **`maximumDepth`** | Maximum depth for the tree. (0 means no limit.) | `0` |
+
+### Training
+
+If training is not done as part of the constructor call, it can be done with the
+`Train()` member function, which has several overloads.
+
+ * `tree.Train(`_`data`_`, `_`labels`_`, `_`numClasses`_`)` (train on numerical-only data)
[…]
