### 2023-01-05 4245467
Added XOR test
