### 2023-01-09 77d9d03
Add -lgfortran and -lquadmath when statically linking OpenBLAS on Linux.
