### 2023-01-09 f81359b
Try to bundle the upstream version of cereal directly.
