+++ .github/workflows/main.yml
@@ -57,7 +57,13 @@ jobs:
         - name: Install Build Dependencies
           run: |
             sudo apt-get update
-            sudo apt-get install -y --allow-unauthenticated libopenblas-dev liblapack-dev g++ libcereal-dev libensmallen-dev libhdf5-dev libarmadillo-dev libcurl4-openssl-dev
+            # We don't install cereal via apt, because the Debian packagers
+            # split the rapidjson dependency into a separate package.  We will
+            # bundle the cereal sources with the R package, so we want them to
+            # be exactly the upstream sources (with rapidjson included).
+            sudo apt-get install -y --allow-unauthenticated libopenblas-dev liblapack-dev g++ libensmallen-dev libhdf5-dev libarmadillo-dev libcurl4-openssl-dev
+            wget https://github.com/USCiLab/cereal/archive/refs/tags/v1.3.2.tar.gz
+            tar -xvzpf v1.3.2.tar.gz
 
         - name: Install R-bindings dependencies
           run: |
@@ -69,7 +75,7 @@ jobs:
         - name: CMake
           run: |
             mkdir build
-            cd build && cmake -DDEBUG=OFF -DPROFILE=OFF -DBUILD_CLI_EXECUTABLES=OFF -DBUILD_PYTHON_BINDINGS=OFF -DBUILD_JULIA_BINDINGS=OFF -DBUILD_GO_BINDINGS=OFF -DBUILD_R_BINDINGS=ON -DDOWNLOAD_DEPENDENCIES=ON -DBUILD_TESTS=ON ..
+            cd build && cmake -DDEBUG=OFF -DPROFILE=OFF -DBUILD_CLI_EXECUTABLES=OFF -DBUILD_PYTHON_BINDINGS=OFF -DBUILD_JULIA_BINDINGS=OFF -DBUILD_GO_BINDINGS=OFF -DBUILD_R_BINDINGS=ON -DDOWNLOAD_DEPENDENCIES=ON -DBUILD_TESTS=ON -DCEREAL_INCLUDE_DIR=../cereal-1.3.2/include/ ..
 
         - name: Build
           run: |
@@ -137,11 +143,7 @@ jobs:
         if: runner.os != 'Windows' && runner.os != 'macOS'
         run: |
             sudo apt-get update
-            sudo apt-get install -y --allow-unauthenticated libcurl4-openssl-dev libcereal-dev
-
-      - name: Install check dependencies
-        if: runner.os == 'macOS'
-        run: brew install cereal
+            sudo apt-get install -y --allow-unauthenticated libcurl4-openssl-dev
 
       - name: Install dependencies
         run: |
