### 2023-01-09 a6f2ee2
Actually perform monochromatic range search, not bichromatic with the same sets.
