### 2023-01-10 8fe1826
Try to work around Visual Studio bug.
