### 2023-01-13 1fc89d5
Check Shape-Size in GMM
