### 2023-01-13 5d29387
Comment out pragmas in base64.hpp to fix CRAN warnings.
