+++ .github/workflows/main.yml
@@ -64,6 +64,10 @@ jobs:
             sudo apt-get install -y --allow-unauthenticated libopenblas-dev liblapack-dev g++ libensmallen-dev libhdf5-dev libarmadillo-dev libcurl4-openssl-dev
             wget https://github.com/USCiLab/cereal/archive/refs/tags/v1.3.2.tar.gz
             tar -xvzpf v1.3.2.tar.gz
+            # These directives cause warnings on CRAN:
+            # https://github.com/USCiLab/cereal/blob/master/include/cereal/external/base64.hpp#L28-L31
+            # The command below comments them out.
+            sed -i 's|#pragma|// #pragma|' cereal-1.3.2/external/base64.hpp
 
         - name: Install R-bindings dependencies
           run: |
