### 2023-01-13 #5026 @rcurtin
> I have finished my initial review. I think overall, mlpack 4 is a well written package that has indicated great potential in the machine learning field, considering its good documentation, user friendliness, and the algorithms it implemented, etc. Thus, I would recommend it for publication after minor revision:
>
>     1. even though I understand that in mlpack's website, there are some comparisons between mlpack and other similar packages, it is not mentioned in the manuscript itself. I would suggest the authors to add such descriptions, based on the requirements of "State of the field" and "performance" in the checklist.

@zhangjy-ge We do have some benchmarks on the homepage of the website.  We can reference them more prominently in the paper---or do you mean to include some benchmarks specifically in the paper itself?  (Personally I think it might be too long with a benchmarking table and sufficient description.)

> @rcurtin, to give a progress update, since I had not used this library before, I was going through the installation process and checking out some of the functionality. I ran into some issues during that step last week, and I will be continuing my review this week.

@sandeep-ps happy to help out, just send a ping if you are having any issues.  (You could even open an issue on mlpack's Github if you wanted, but whatever way is fine.)  Often it can be easier to install mlpack via the system package manager, but our recent refactorings to make mlpack header-only have made the process much easier than it used to be.
