### 2023-01-14 #5026 @zhangjy-ge
@rcurtin After double check, it seems a performance comparison is optional but we should still explicitly state the difference between mlpack and other packages (from `Do the authors describe how this software compares to other commonly-used packages?`) So maybe a short description of mlpack's differences comparing to other packages will be helpful.
