### 2023-01-17 d9a32f4
Add random splitting for categorical features for decision trees.
