+++ src/mlpack/methods/decision_tree/decision_tree.hpp
@@ -24,6 +24,7 @@
 #include "random_binary_numeric_split.hpp"
 
 #include "all_categorical_split.hpp"
+#include "random_categorical_split.hpp"
 
 #include "all_dimension_select.hpp"
 #include "random_dimension_select.hpp"
