### 2023-01-22 8731672
Added param check in Logistic regression.
