### 2023-01-22 mlpack/mlpack#3367 @sandeep-ps
The compilation goes away after changing `RandomForest` to `RandomForest<>`, as mentioned in https://github.com/mlpack/mlpack/pull/3341. I'm pretty sure, I'm using mlpack 4.0.0. So, it's clear where the problem lies.
