### 2023-01-22 mlpack/mlpack#3368 @sandeep-ps
A statement of need missing from the documentation
#### Problem location

https://www.mlpack.org/docs.html

#### Description of problem

This issue is related to the JOSS paper review https://github.com/openjournals/joss-reviews/issues/5026 and the review criteria checklist entry:
`Documentation - A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?`

A statement of need or a similar section seems to be missing from the documentation (https://www.mlpack.org/docs.html), and it needs to be added to meet the review criteria.

A related and potentially minor item. I could not find the target audience explicitly stated in the documentation. Even though it's clear to me that this library is for software developers that use machine learning, it needs to be documented based on the review criteria. Also, if you would like to elaborate more on the target audience or if there is a more specific audience that you may be targeting (e.g., beginner vs. advanced developer), it would be good to mention that in the software paper and the software's documentation.
