### 2023-01-22 mlpack/mlpack#3369 @sandeep-ps
C++ API documentation question
This question is related to the JOSS paper review https://github.com/openjournals/joss-reviews/issues/5026 and criteria checklist entry `Documentation - Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?`

Since this is primarily a C++ library, are the library's C++ API methods documented somewhere? I could see similar documentation for other programming language bindings (e.g., https://www.mlpack.org/doc/stable/python_documentation.html), but not the ones for C++.
