### 2023-01-22 mlpack/mlpack#3370 @sandeep-ps
Software paper authorship criteria question
This question is related to the JOSS paper review https://github.com/openjournals/joss-reviews/issues/5026 and the checklist entry `General checks - Contribution and authorship: Has the submitting author (@rcurtin) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?`

It's easy to determine that `@rcurtin` has made significant contributions to the software by looking at the git commits. To evaluate the second question above, I need to understand how the other co-authors were determined. Could you please share the authorship criteria for the remaining authors?

Also, https://www.mlpack.org/community.html doesn't contain all authors listed in the paper. Could that list be updated to include all authors in the software paper?
