### 2023-01-22 mlpack/mlpack#3371 @sandeep-ps
Statement of need section in the software paper missing discussion on related libraries
This issue is related to the JOSS paper review https://github.com/openjournals/joss-reviews/issues/5026 and checklist entry: `Software paper - A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?`

This issue is also related to https://github.com/mlpack/mlpack/issues/3368 and https://github.com/openjournals/joss-reviews/issues/5026#issuecomment-1382659473.

The current `Statement of need` section does not discuss how mlpack relates to other similar machine learning C++ libraries. Please add text explaining this aspect.

 Also, please add any discussion on the target audience to the paper's `Statement of need` section. Thanks.
