### 2023-01-22 #5026 @sandeep-ps
I have also completed my initial review. In general, mlpack 4.0.0 is a well-written machine-learning library with good documentation, examples, community guidelines, etc.

Apart from zhangjy-ge's comment above, which I agree with, I have opened a few GitHub issues in the mlpack repository for further clarification and adding some missing content. I'm looking forward to more discussion on individual issues. Thanks.
