+++ doc/quickstart/cpp.md
@@ -105,11 +105,11 @@ int main()
       testLabels, 0.3);
 
   // Create the RandomForest object and train it on the training data.
-  RandomForest r(trainDataset,
-                 trainLabels,
-                 7 /* number of classes */,
-                 10 /* number of trees */,
-                 3 /* minimum leaf size */);
+  RandomForest<> r(trainDataset,
+                   trainLabels,
+                   7 /* number of classes */,
+                   10 /* number of trees */,
+                   3 /* minimum leaf size */);
 
   // Compute and print the training error.
   Row<size_t> trainPredictions;
