### 2023-01-24 mlpack/mlpack#3367 @rcurtin
Thanks @sandeep-ps!  You are indeed correct.  This is just a typo in the quickstart.  I opened PR #3373 to fix the issue. :+1:
