### 2023-01-24 mlpack/mlpack#3368 @rcurtin
Thanks for pointing this out.  I updated the website: https://www.mlpack.org/docs.html in this commit: https://github.com/mlpack/mlpack.org/commit/58ae7214ce550bf1bbec2eacd3d68094d5d1a8a1 and now the issue should be handled. :+1:  I think you are right that some additional clarification could be done in the paper, and so we will make those changes also.
