### 2023-01-24 mlpack/mlpack#3370 @rcurtin
I asked people on the Contributors team if they were interested in being a part of the paper.  To be on the Contributors team, someone has to have contributed significantly to the effort and be nominated to the team as per the process in the [governance document](https://github.com/mlpack/mlpack/blob/master/GOVERNANCE.md).  It's a contributor's choice to add themselves to the list of authors/maintainers, and some people on the paper have chosen to not do that. :+1:
