### 2023-01-24 2e4f8bf
Add target audience and comparison with other libraries.
