+++ paper/paper.bib
@@ -210,3 +210,37 @@ and Applications: Software For Humanity},
     url = {https://CRAN.R-project.org/package=mlpack},
   }
 
+@article{johnson2019billion,
+  title={Billion-scale similarity search with {GPUs}},
+  author={Johnson, Jeff and Douze, Matthijs and J{\'e}gou, Herv{\'e}},
+  journal={IEEE Transactions on Big Data},
+  volume={7},
+  number={3},
+  pages={535--547},
+  year={2019},
+  publisher={IEEE}
+}
+
+@inproceedings{muja2009fast,
+  title={Fast approximate nearest neighbors with automatic algorithm configuration},
+  author={Muja, Marius and Lowe, David G},
+  booktitle={Proceedings of the 2009 International Conference on Computer Vision Theory and Applications (VISAPP '09)},
+  year={2009},
+  pages={331--340}
+}
+
+@article{dlib09,
+  author={Davis E. King},
+  title={Dlib-ml: A Machine Learning Toolkit},
+  journal={Journal of Machine Learning Research},
+  year={2009},
+  volume={10},
+  pages={1755--1758}
+}
+
+@book{alexandrescu2001modern,
+  title={Modern {C}++ {D}esign: generic programming and design patterns applied},
+  author={Alexandrescu, Andrei},
+  year={2001},
+  publisher={Addison-Wesley}
+}
