+++ HISTORY.md
@@ -1,5 +1,6 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+  * Fix a few missing includes in `<mlpack.hpp>` (#3374).
 
 ### mlpack 4.0.1
 ###### 2022-12-23
