### 2023-01-25 ad2325f
Update COPYRIGHT.txt
