### 2023-01-25 ed882bd
Update COPYRIGHT.txt
