### 2023-01-25 8e6df70
Changed Armadillo to explicit version detection.
I am proposing a change to the armadillo's version detection. 
A more explicit version detection has been updated in this header file
