+++ src/mlpack/core.hpp
@@ -61,8 +61,8 @@
   #include <omp.h>
 #endif
 
-// Use Armadillo's C++ version detection.
-#ifdef ARMA_USE_CXX11
+// C++ version detection.
+#if __cplusplus >= 201103L
   #define MLPACK_USE_CX11
 #endif
 
