### 2023-01-26 mlpack/mlpack#3367 @conradsnicta
the fix in PR #3373 has been merged
