### 2023-01-26 9b6d770
remove unused define
