+++ src/mlpack/bindings/julia/julia_util.h
@@ -4,6 +4,11 @@
  *
  * Some utility functions in C that can be called from Julia with ccall() in
  * order to interact with the IO interface.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
 #ifndef MLPACK_BINDINGS_JULIA_JULIA_UTIL_H
 #define MLPACK_BINDINGS_JULIA_JULIA_UTIL_H
