+++ src/mlpack/bindings/julia/generate_jl.cpp.in
@@ -3,6 +3,11 @@
  * @author Ryan Curtin
  *
  * This is a template file to call the PrintJL() function for a given binding.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
 #define BINDING_TYPE BINDING_TYPE_JL
 // Disable debug output.
