+++ src/mlpack/bindings/R/generate_R.cpp.in
@@ -3,6 +3,11 @@
  * @author Yashwant Singh Parihar
  *
  * This is a template file to call the PrintR() function for a given binding.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
 #define BINDING_TYPE BINDING_TYPE_R
 // Disable debug output.
