### 2023-01-26 mlpack/mlpack#3371 @rcurtin
Thanks for pointing this out---we added some text to compare mlpack to some other libraries.  I'll comment more in the review issue, since the other reviewer wanted to see that improved too.  (I'll also get the bot to rerender the paper for easier inspection of changes.) :+1:
