### 2023-01-26 #5026 @rcurtin
Thanks for all of the comments @sandeep-ps and @zhangjy-ge!  We have addressed each of them by resolving the various issues in the mlpack repository, and we added a paragraph to the `Statement of Need` section comparing mlpack with other C++ machine learning libraries. :+1:
