+++ README.md
@@ -107,8 +107,7 @@ Citations are beneficial for the growth and improvement of mlpack.
  - C++14 compiler
  - [Armadillo](https://arma.sourceforge.net/docs.html)      &nbsp;&emsp;>= 9.800
  - [ensmallen](https://ensmallen.org)      &emsp;>= 2.10.0
- - [cereal](http://uscilab.github.io/cereal/)         &ensp;&nbsp;
-&emsp;&emsp;>= 1.1.2
+ - [cereal](http://uscilab.github.io/cereal/)         &ensp;&nbsp;&emsp;&emsp;>= 1.1.2
 
 If the STB library headers are available, image loading support will be
 available.
