### 2023-01-27 194f629
Update README.md
Improving and moving instructions for cmake error
