+++ src/mlpack/bindings/julia/tests/runtests.jl
@@ -2,6 +2,12 @@
 # @author Ryan Curtin
 #
 # Tests for the Julia bindings.
+
+# mlpack is free software; you may redistribute it and/or modify it under the
+# terms of the 3-clause BSD license.  You should have received a copy of the
+# 3-clause BSD license along with mlpack.  If not, see
+# http://www.opensource.org/licenses/BSD-3-Clause for more information.
+
 using Pkg
 Pkg.activate(".")
 using Test
