### 2023-01-29 #5026 @rcurtin
Ah!  The semicolon is the trick.  I was trying to play with that earlier and didn't figure out how.  Clearly I did not read the example paper closely enough, since it's documented quite clearly.

Anyway, I pushed fixes for all instances where this happened.  I'll ask editorialbot to build again. :+1:
