+++ README.md
@@ -136,8 +136,8 @@ If the `cmake ..` command fails due to unavailable dependencies, consider either
 are installed, e.g. using the system package manager.  For example, on Debian and Ubuntu,
 all relevant dependencies can be installed with `sudo apt-get install libarmadillo-dev libensmallen-dev libcereal-dev g++ cmake`.
 
-Alternatively, since CMake v3.14.0 the `cmake` command can create the build folder itself.
-The above commands can be rewritten as follows:
+Alternatively, since CMake v3.14.0 the `cmake` command can create the build folder itself,
+the above commands can be rewritten as follows:
 
 ```sh
 cmake -S . -B build
