+++ paper/paper.bib
@@ -15,7 +15,8 @@
   volume={3},
   number={26},
   pages={726},
-  year={2018}
+  year={2018},
+  doi={10.21105/joss.00726}
 }
 
 @article{jordan2015machine,
@@ -26,7 +27,8 @@
   number={6245},
   pages={255--260},
   year={2015},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.aaa8415}
 }
 
 @article{carleo2019machine,
@@ -37,7 +39,8 @@
   number={4},
   pages={045002},
   year={2019},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/RevModPhys.91.045002}
 }
 
 @article{devlin2018bert,
@@ -52,7 +55,8 @@
   author={Gieseke, Fabian and Igel, Christian},
   booktitle={ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
   pages={1445--1454},
-  year={2018}
+  year={2018},
+  doi={10.1145/3219819.3220124}
 }
 
 @article{paleyes2020challenges,
@@ -60,7 +64,8 @@
   author={Paleyes, Andrei and Urma, Raoul-Gabriel and Lawrence, Neil D},
   journal={ACM Computing Surveys (CSUR)},
   year={2020},
-  publisher={ACM New York, NY}
+  publisher={ACM New York, NY},
+  doi={10.1145/3533378}
 }
 
 @article{lavin2022technology,
@@ -71,7 +76,8 @@
   number={1},
   pages={1--19},
   year={2022},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.21203/rs.3.rs-133138/v1}
 }
 
 @article{fischer2020ai,
@@ -82,7 +88,8 @@
   number={1},
   pages={56--83},
   year={2020},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/make3010004}
 }
 
 @misc{mlpack2021vision,
@@ -98,7 +105,8 @@
   booktitle={International Conference on Similarity Search and Applications (SISAP 2016)},
   pages={221--235},
   year={2016},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-46759-7_17}
 }
 
 @inproceedings{curtin2017dual,
@@ -107,7 +115,8 @@
   booktitle={SIAM International Conference on Data Mining (SDM '17)},
   pages={300--308},
   year={2017},
-  organization={SIAM}
+  organization={SIAM},
+  doi={10.1137/1.9781611974973.34}
 }
 
 @article{lee2008fast,
@@ -127,7 +136,8 @@
   number={4},
   pages={229--253},
   year={2014},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/sam.11218}
 }
 
 @inproceedings{fang2016m3,
@@ -152,7 +162,8 @@
   volume={1},
   number={2},
   pages={26},
-  year={2016}
+  year={2016},
+  doi={10.21105/joss.00026}
 }
 
 @article{curtin2021ensmallen,
@@ -169,7 +180,8 @@
   author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
   booktitle={Proceedings of the 20th International Conference on Electronic Publishing},
   year={2016},
-  pages={87--90}
+  pages={87--90},
[…]
