### 2023-02-01 094d9b6
Update doc/user/formats.md
Co-authored-by: Ryan Curtin <ryan@ratml.org>
