### 2023-02-01 54ec304
fixing typo in trees.md
