### 2023-02-01 53909b3
fixing typo in range_search.md
