+++ doc/tutorials/range_search.md
@@ -3,7 +3,7 @@
 Range search is a simple machine learning task which aims to find all the
 neighbors of a point that fall into a certain range of distances.  In this
 setting, we have a *query* and a *reference* dataset.  Given a certain range,
-for each point in the *query* dataset, we wish to know all points in the \b
+for each point in the *query* dataset, we wish to know all points in the
 reference dataset which have distances within that given range to the given
 query point.
 
