### 2023-02-01 fa73dbb
fixing links in approx_kfn.md
