### 2023-02-01 c37d9f7
fixing typo in approx_kfn.md
